requirements.txt
langchain-openai
matplotlib
notebook
pandas
python-dotenv
scikit-learn
